requirements.txt
python-memo
nltk
pandas